refactor: update creativity constants file
